requirements.txt
absl-py==0.9.0
asgiref==3.2.10
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
Django==3.1
djangorestframework==3.11.1
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.1
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
joblib==0.16.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2
nltk==3.5
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
protobuf==3.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
tensorboard==2.1.0
tensorflow-cpu==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tqdm==4.48.2
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2